Add a persist_mapping option to the mapping file
This adds the option to permanently rewrite the configuration so the
aliases are used instead of the system name.  This is useful where
you have a variety of hardware and you want to have consistent device
naming accross all platforms - this allows you to essentially rename
the interfaces permanently so they match the abstracted nicN names.

Note, this needs to be run with --cleanup or the old (now conflicting)
configs will still be in place, and it may require a reboot before
the changes are fully applied.

Change-Id: I5af146e764b72c4beaa41c549fabff0af8802152