Set primary interface on OVS bonds
This change sets one of the member interfaces of a bond as the primary
interface, which results in that interface being the active slave.
This change adds a step to the apply method in impl_ifcfg which runs
'ovs-appctl bond/set-active-slave <bond> <iface>' after bringing up
the bond interfaces. This step ensures that the bonds work correctly
without LACP switch support. If a member interface on the bond is set
as primary, that interface will be used.

Co-Authored-By: Dan Prince <dprince@redhat.com>

Change-Id: I795bb3b8ef977f9276bfec062b197c473393942e