Automate selection of active bond slave
This change selects an interface to be the active slave in a bond
with active/passive characteristics. If one of the interfaces is
marked as primary, it will be the active slave. If none of the
interfaces are marked, the interface with the lowest alphanumeric
value will be chosen. For instance em2 comes before em3.

Co-Authored-By: Dan Prince <dprince@redhat.com>

Change-Id: Ic9b4e8c68b788b98a19ea33a76c9210a80deabeb