os-net-config: ensure ifup is called just once
This patch fixes an issue in the ifcfg provider implementation
that caused interfaces to be started twice. This can
cause failures if the interface uses DHCP and dhclient
was already running.

The child_members function was incorrectly adding the interface
name to the child members set on key error. This patch removes
that logic and simply ignores key errors.

Change-Id: Ibe0e32bc09979bc68b92a722b2bfa383e77502a9
Closes-bug: #1485634