Add ability to set IPv6 routes in os-net-config templates
This change adds the ability to write IPv6 routes to the system
NIC configuration. In the ifcfg provider, IPv6 routes must be
written to network-scripts/route6-<interface> rather than
route-interface. This patch checks the next hop, and if it is an
IPv6 address (assumed because of the presence of ":"), then the
route will be written to route6-<interface>. This patch includes
tests to confirm proper operation.

Change-Id: I3563800c774a6082a896c7508f5a92e25418fe63