launchpad bug 1537330, fix _is_active_nic
The detailed bug description is at
https://bugs.launchpad.net/os-net-config/+bug/1537330

Since kilo, os-net-config starts to support linux bond.
However, the addr_assign_type value of an active member
interface of a linux bond is 3, but not 0. The consequence
of this problem is that the linux bond member links are
never considered as active, this further messes up the
nicX to interface name mapping.

This change checks if device directory exists, instead
of checking the addr_assign_type value. This directory
exists only when the interface is an actual physical
interface.

Change-Id: I8e4c95a2efa809fd236b07cbce1b81a2e774f858