Fix hierarchy for Linux Bonds and Linux Bridges
This change cleans up some of the Linux Bond and Linux Bridge logic
that processes member interfaces and VLANs.

One bug is fixed, where a VLAN might be assigned as a slave
interface on a bond.

Changes were made so that if VLANs are placed under a Linux Bond,
the PHYSDEV is set to the bond, so the device: does not need to
be specified in the config.

This change ensures that if a bridge has a bond and VLANs as members,
that the VLAN will be a member of the bridge and not the bond.

Change-Id: Id329737f0032e781b111741086ded83f378297a7