Fix order-of-operations bug in os-net-config restart_interfaces
The current order of operations of bringing up interfaces starts
with the bridges, and restarts the members of the bridges. This
works well when an OVS bond is part of an OVS bridge, but does not
work well when Linux bonding is used with no bridges (such as in
the case of the Nuage plugin, which creates the bridges outside of
the ifcfg files).

This change will ensure that Linux bonds are brought up first, then
any bridges are brought up, interfaces are brought up next, and the
VLAN interfacs are brought up last. This corrects a race condition
where any VLAN which was brought up before the bond would not be
active, and the VLAN would have to be brought up by hand.

This change also fixes a logging issue where os-net-config would
report that no changes were necessary when changes were made. Now,
the logging messages indicating that "No changes were necessary"
show up when no changes are required, instead of when changes are
actually made.

Change-Id: I1efee3dfd8e8cef01b054bb57a3085cc7eb60372