Enhance and fix NIC mapping feature
The original implementation supported a strictly numeric mapping scheme
(e.g. nic1, nic2) that could misbehave if an active NIC was not listed in the
user's mapping file. This change fixes the misbehavior, and enhances the
feature by not requiring NIC aliases follow the numeric mapping scheme. This
allows the user to choose meaningful names for the NIC aliases.

NIC mapping now happens in two steps:

1) Process any user supplied mappings
   - NIC alias does not need to follow the numeric "nicN" scheme
   - Existing validation rules apply: mappings for inactive NICs are ignored
     (but logged), and duplicate mappings are rejected
2) Generate default mappings as needed
   - Only applies to active NICs that aren't already mapped
   - Follows the numeric scheme (nic1, nic2) using the NIC number based on
     the list of active NICs
   - No default mapping is assigned if another NIC is already using that
     numeric alias

Change-Id: I6943623a51702349f6a7dcf2de8a8429078a3ab0
Closes-Bug: 1612723