Fix Linux bond/slave ifup ordering
The latest version of os-net-config is not reliably enabling bonding
when there are no existing bonds on the system. This is due to an
order-of-operations problem where it tries to bring up the bond
before the slave interfaces. This can result in an error because
the bonding module is not present.

This change activates the bond only after the interfaces have been
activated. This should ensure that the bonding module is loaded.

Change-Id: I62e4b10a88168948ed59170285223fb4f8d8de88
Fixes-bug: 1611471