Remove child members activation for OVS-DPDK bond
While creating DPDK bond, if-up of the bond interface
itself will activate the member interfaces. Individual
members should not be called for ifup.

Additionally, removing the option to select the primary
active-slave member for the bond, because of #1654975.
Once it is fixed, primary selection from the network
config will be added.

Change-Id: I83527ada4a3618927d55b7aa06d436686b12d5b8
Closes-Bug: #1643026.