Allow setting NM_CONTROLLED=yes in os-net-config ifcfg files
This change adds a flag to the base interface type, nm_controlled.
If this flag is true, the ifcfg file will contain NM_CONTROLLED=yes.
Since this flag is applied at the base it applies to any interface
type.

Note that not all interface types are supported by NetworkManager
at this time, so this option should be used with caution.

A demonstration of the usage is included in the
linux_bond_networkmanager.yaml file in the sample directory.

Change-Id: I2df6ce5b4bdb04651f27fc5daa64aa752e47f3b1
Closes-Bug: 1671888