os_net_config should map nics that are down if nic is in mapping file
Currently os-net_config will map nics from a user-supplied mapping file
only if the nic is active (operstate = up).  This can cause problems
if a nic is in a bond and one of the bond's nics has no carrier.
This fix will map the nic from the mapping file if the nic is defined
on the system, regardless of the operstate status.

The fix implements a new function to return a list of available nics
(no check of operstate) for use if a mapping file is supplied.  The
list of active nics must still be used in the default case when
numbering nics (no mapping file supplied).  There is also some
cleanup to check if a user-supplied mac is in the mapping file before
attempting to convert the mac to a nic.

Change-Id: Ia5d8c8b49b7ac0b51ee42a754f06e5e53587a5f6
Closes-Bug: 1679787
(cherry picked from commit b905e0d803e55deee9620a24de4b87c3429b43ab)