Configure mtu value for DPDK port
MTU value is already written to ifcfg file if provided on the interface,
but for dpdk port, we have to apply this to ovs-vsctl command via ovs_extra.
This patch is adding the support to configure the given mtu vlaue to the
dpdk port using ovs_extra.

implements: blueprint ovs-2-6-features-dpdk

Change-Id: Ic01ed8cee2edbd648de0b64dc7f01da80c153d81