Only log that mapping file is in use if it's actually true
os-net-config always logs the message "Using mapping file at:
/etc/os-net-config/mapping.yaml" even if the file does not exist, in
which case it's not actually being used at all. Move the log message to
only be shown if the mapping file actually exists and is used.

In the case where no mapping file is used, also log a message indicating
such.

Change-Id: Iea46e215d531e0fed5839c28c116417db464d52d
Closes-Bug: #1737191