Use INFO log level for getting nics
Getting the ordered, active, and available nics has been traditionally
one of the parts of os-net-config that can cause issues due to either
user error or bugs. There are helpful log messages to log what is going
on in the logic, but it's all hidden behind the DEBUG log level.

These messages should be INFO instead of DEBUG. It's much more user
friendly when there are issues instead of having to rerun with --debug.

Change-Id: I7d9fd87bcbba87ecef4fedcb2b6e8dd48c3e610e
Closes-Bug: #1737196