Restore the order of params in ifcfg file that was inadvertently changed
This change to os-net-config in Queens changed the order of the ONBOOT
parameter in the ifcg files created by os-net-config -
https://review.openstack.org/#/c/485132/.  When upgrading a diff is done
between the existing ifcfg files and new data to write.  Since the order
is different, the ifcg files will be detected as different and interfaces,
bridges, etc. will be restarted.  Change the order back to avoid
this from happening.

Change-Id: I77162d28b1dc173e3a90cb385a3af92e21c71574