Fix numbered NIC mapping when using dotted VLAN notation
When declaring VLANs using dotted notation (such as em1.10
for VLAN 10 on interface em1), numbered NICs are not
properly mapped to real interface names. This patch fixes
the handling of the mapped NIC names inside the BaseOpts
object that is the base class for all interface types.
This allows an interface name of "nic1.10" to be mapped
to, e.g. "em1.10"

Closes-bug: #1783206

Change-Id: Ibd9b28128900d9e959b82eb8673b8380bf20606a
(cherry picked from commit e12837353dc3668da7cfc575777ef197df4b0497)