Apply IP/netmask/route/MTU changes without bouncing interfaces.
Modifications to impl_ifcfg.py to support making changes to IP
address, MTU, and routes live, rather than running ifdown/ifup on the
interface. Bouncing the interface is very disruptive, especially with
bridges where all the child interfaces are also bounced.

If only the IP address, netmask, MTU, routes, and/or boot status are
changed, the file will be written and 'ip' or 'ip route' commands
will be run to apply the new config on the fly. If any commands fail,
the interface will be restarted instead.

Closes-bug: 1795129
Change-Id: I75239fd3f273a88feb80f09a67aa5a947d64ac30