Fix - restart VLAN interface on underlying device change
When a change is needed on bond, os-net-config write the
configuration and restart nic and bond. But it does not
restart vlan on top of the bond. The results is that  all
routes for vlan interfaces are lost.

This moves the processing of VLAN's so that it run's
after any interfaces, bridges, bonds etc. Then
concatenates the restart lists from other interfaces and
do a check if 'PHYSDEV' of a VLAN interface is being
restarted.

Closes-Bug: #1806027
Related: RHBZ#1654987
Change-Id: I246da54cf02d06466f52d210f89d82ca9e2a6ef2
(cherry picked from commit 19067b4e77ec618ffef67acc771869ecbc33ad34)