Add Policy Based Routing Support to os-net-config
Add support for routing tables and rules, as well as a
property to specify a routing table for each route to
the impl_ifcfg implementation.

This change adds a new top-level object in the
network_config with type "route_table" to specify
which route table IDs and names should be added to
the /etc/iproute2/rt_table file.

This change adds a "table" property to routes for each
interface type, that allows the selection of which
routing table to apply the route to.

This change also adds a "rules" property to each
interface type, which can be used to specify a list
of rules. Each rule contains a rule string and an
optional comment, both of which will be added to the
/etc/sysconfig/network-scripts/rule-<iface> file.

This change includes tests and changes to the schema
for validation.

Note that this change was based on the abandoned patch
https://review.openstack.org/#/c/575712
and was recreated when that patch diverged too far
from master branch.

Change-Id: I6906d3b6923845af177faba6579fa255a2195bec
Closes-bug: #1783297
(cherry picked from commit 35832347e1d14b0cd8c3e58ab7d4a5a86f6716f6)