Fix os-net-config interface restarts when applying routes
There is a bug in the os-net-config code that applies updates to
existing routes. Instead of deleting routes that no longer exist
in the config and adding routes on the fly that do, the process
always fails so the interface gets restarted.

This change fixes the external call to "/sbin/ip", but also
fixes a bug where the name of the file containing the stored
routes was being passed incorrectly. This caused os-net-config
to always think there were no routes present, and to apply the
new route without deleting the old route. This can lead to an
error if the routes conflict.

Change-Id: I4315e9812c641a667a1d5c6529cdba5d2f5bf640
Closes-bug: 1819212
(cherry picked from commit e912b02c3bec61ec3af610820601cc24b5dc1cba)