Add ovs_options and ovs_extra parameters to VLAN object
This patch adds the ovs_options and ovs_extra parameters to the
VLAN object, and OVS_OPTIONS/OVS_EXTRA to the OVS VLAN ifcfg.
An OVS VLAN is an internal port on the bridge, and when ifup
is run the ovs-vsctl command will be called with the commands
in OVS_OPTIONS and the list of commands in OVS_EXTRA. This
makes OVS VLANs work the same as other OVS object types.
These settings are ignored for VLANs not on an OVS bridge.

Change-Id: Ie1db61eea48e3426a59ae02b13dd8ab8ef85e851
Closes-bug: 1821816