Update docs testing
Use doc/requirements for documentation requirements, this avoids pulling
in sphinx into the python environments and thus solves the problem that triggered
https://review.opendev.org/658224 as well.

Note that both changes are needed in general, while the problem is solved
with either.

Switch to openstackdocstheme instead of oslosphinx, update sphinx
requirements for this and sync with global requirements.

Change-Id: I360c185af8740a71ebc0f5f49debd5b6032161ec