ifdown does not work if no ifcfg file exist
ifdown happen before writing_config and if interface is up by
dhcp/dhclient it will not be down, ifdown relaying on ifcfg exist.

causing issue when pxe boot interface is also used as part of bond
interface.
interface is required to be down before adding to bond,
this patch validate that any physical interface will be in down state.

Change-Id: Ie096bf1296062744193bdcb9713ee42405e206d8
Closes-Bug: #1836024