fix missing shared_block for PF interface in switchdev bond connected to ovs
When having a deployment with ovs bridge having a linux bond and
the slaves of that bond are in switchdev mode , shared_blocks are
expected to be created for both the bond and the slaves, but because the
slaves have an autoconfig IPv6 so ovs will assume that the slaves are not
managed by it, and so those shared_blocks will be missing for those slaves.

As a workaround for this issue we should handle two cases:
 * First os-net-config run / Post deployment
 * Reboot compute node

For deployment case, we are restarting openvswitch and ifdown ifup
any PF interface, this will make openvswitch recreate PF shared_block

For reboot case, we are disabling ipv6 for PF interfaces and because
the ovs will start after moving to swichdev, so the shared_blocks will
be created.

Also there is no need to ifup the PF interfaces after moving to switchdev
in reboot case as the config scripts will bring them up later

Change-Id: I2468866629d95dc9f088985d57fbf70263253235
Closes-Bug: #1836581
(cherry picked from commit a74bbdd7b017fb3bb9faffb4240efa59cb42c79c)