Adding IPoIB child interfaces to the os-net-config
This patch adds support for implementing IPoIB child interfaces using
the os-net-config. It allows the user to create persistent IB childs
using ifcfg implementation. It adds the type ib_child_interface to the
os-net-config interfaces which needs the parent interface and the pkey
id as required fields. The IPoIB child interface should be a seperate
interface from the parent interface.

Change-Id: I9ae739c892318c28ad6165a96644719fa1aa693d