Enable new routing rules on the fly without restarting interfaces
This change enables os-net-config to add and delete routing rules
without restarting interfaces. The iproute2 binary is called for
each rule that is added or deleted, and the updates happen without
the interface being restarted. The config files will be updated so
that on the next interface restart the changes will be permanent.
This change fixes bug 1865123.

Change-Id: I83f8b26000148a9b311ddf43cd423475373be404
Closes-bug: 1865123