Run ifup on a bond when a slave interface is restarted
This change adds the bond to the list of interfaces to be
restarted when a slave interface is restarted. On Fedora/CentOS
like systems restart all interface slaves in a bond, the bond
will be left in a downed state. This change runs ifup on the
bond if any slave interfaces are restarted, which insures that
the bond should be up in all cases when os-net-config exits.
Tests were added to ensure that in all cases each interface
or bond will only be restarted once, and that bonds will only
be restarted if the interface modifications require restart.

Change-Id: Ibbe8021afdc4772c6e2017d11d012b8f5c46d907
Closes-bug: 1870608
(cherry picked from commit fe13c4944be3221092f9147385700e4914a30cde)