Skip VFs configured for DPDKs during NIC ordering
During NIC ordering all the VFs are expected to be skipped.
But if DPDK drivers are bound to a VF, then those VFs are still taken
in to account and is affecting the NIC ordering. The VFs bound to DPDK
drivers also shall be removed from the NIC ordering logic resulting in
consistency even with the re-run of os-net-config.

Change-Id: I50af9f75b01addae0e13830cc7a863072f9be34c