Avoid openvswitch restarts during re-run
For offload configuration on the first time, opevswitch
need to be restarted after the interface configured to
switchdev mode to enable VF-LAG. But once it is configured,
openvswitch should not be restarted, which will create
VF-LAG failures. In general, in the TripleO during
updates/upgrades, ovs restart will not be triggered
and it will be allowed to the user to restart or reboot
the node. Aligning with it, opevswitch restart has been
avoided on all os-net-config re-runs.
Closes-Bug: #1887957

Change-Id: I0a3b374b321ba039c8e0b1c126ac3eb59e6dc980