Enable smfs software steering when switchdev is enabled
Before this point, when switchdev is enabled, the software
steering mode is not taken into consideration. This patch
tries to enable the smfs software steering mode when the
switchdev mode is enabled, and if it fails, a warning message
would be printed indicating the failur of the proccess, and
the software steering mode will not change from its current value.

Change-Id: I2cee0a5509db7a1c13c9baa056d9d75d16acb0ed
(cherry picked from commit c9b8691f05b35e00a9fec9f013ed0db4942a36f4)