Fixing the dependency for sriov_config service
Some of the 3rd party vswitch providers does not follow the regular ovs db and
vswitchd is split and it is implementation specific. But in general the openvswitch.service
will be used as base even for the 3rd party providers. Moving the dependency from
vswitchd to openvswitch make sriov_config.service to work with other providers.

Closes-Bug: #1903550

Change-Id: Id23051e12dad8d4b84df1821e670661a0891e780