Add mac-size-table to all OVS bridges
The default mac table size may not be sufficient for the
provider bridges. If the mac table overflows, it has been
observed that ovs-vswitchd clocks 100% CPU. So this patch
configures the mac table size of the provider bridges.

The patch does similar to what
https://opendev.org/openstack/puppet-ovn/commit/88aa5c9184f8f448aed499413cda51188d66c7a6
implements in puppet. However, the puppet changes are not
persistent across reboots, therefore it's better to save it
in the OVS database.

Change-Id: Ic0334aaad83bd94b604fbd84fcc21d3520e9d400
Related-bug: #1779706
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>