Exclude DPDK mapped VF device even if not present
As part of FFU from queens to train, os-net-config
will run after the RHEL update. At this point, DPDK
mapping file will have VF if it has been used as
part of NIC partitioning feature, but the physically
the VF will not be present as it is yet to be created.
During the active nic list creation (before creating
VFs), DPDK mapped VF will be used as regular nic,
which results in wrong nic numbering.

This patch excludes the VFs by checking the device
type in the sriov_config.yaml mapping file, so that
the nic numbering will be kept same.

Change-Id: I17d19715a1f7c231630416d6fea06fe9afe44e9b