Check to ensure active DPDK NICs only mapped once
This change ensures that active DPDK NICs are only added
to the list of mapped NICs once, and avoids duplicate entries.

When os-net-config runs for the first time there is no entry
in /sys/net for the DPDK NICs, so we look in
/var/lib/os-net-config/dpdk_mapping.yaml for the list of
DPDK NICs and add them to the NIC mapping list.

If os-net-config has been run before and the DPDK NICs are
already active, there is an entry in /sys/net and also an
entry in /var/lib/os-net-config/dpdk_mapping.yaml for the NIC.
This results in the same NIC being added to the list of active
NICs twice.

Change-Id: Idbee49ce02d000043d0fa8e398aeb82f032c32aa
Closes-Bug: #1912409
(cherry picked from commit ba93c36044a271b52fedb8a275f6b8e05fbe2a62)