Refresh README.rst, expand usage.rst
As part of an effort to fully document os-net-config, this change
makes the README.rst more consistent with other projects, and expands
the usage.rst page to describe useful command-line arguments.

Now that the README has a link to the generated documentation[1], the
provider section can be replaced by the one in the usage.rst page.

[1] https://docs.openstack.org/os-net-config/latest/

Change-Id: I065da7c7028bf081fb1d87ef64c4f3f873bfe111