Document os-net-config YAML format
This change documents the types and attributes used by the TripleO
networking templates[1] along with how these are used by the ifcfg
backend.

The next change in this series will be a new page of complex examples
to illustrate the usage of these types and attributes.

Future changes can document unused types and attributes, and also add
ENI backend notes to go alongside the ifcfg notes.

[1] https://opendev.org/openstack/tripleo-ansible/src/branch/master/tripleo_ansible/roles/tripleo_network_config/templates
Change-Id: Ic9d5643354b76165261498d875266c68e9bc6a7e

Change-Id: Iff3cdb2657b3e427f9fa3a882479253bf2e27fee