Updating sriov_config service to run after openibd service
When having sriov configured in mlnx interfaces and we are using MLNX
ofed, it will install openibd and in reboot case we'll have a race between
sriov_config service and that openibd service which loads mlx drivers,
sometimes it causes failures in starting sriov_config service.
So updating sriov_config service to run after openibd service if it's
present, if not as in default case, systemd ignores it.

Closes-Bug: #1918255
Change-Id: I047e3c40815b71de10e6641c1d39b29de563d8d2