Updating sriov_config.service to run before driverctl.slice and network-pre
On a start/reboot, os-net sriov_config.service
is run to create VFs for nic-partitioned devices.
If vfio-pci driver is bound to any of the VFs,
the VF initialisation doesn't occur properly.
The VF creation has to be completed before driverctl vfio-pci
binding AND all network interface configs.
So modified the sriov_config dependency, to ensure it
is execute before these services

Change-Id: Id0ef377b92c771b62fae99eaedc778ae64b0e52d