Adding logging to file
It's important to support logging to a file for troubleshooting purpose.
It's also important to rotate this file as it might grow quite large
over time.

Also, creating a common module where we can move gradually functions
that are common to any other modules. We currently can't import utils
from sriov_config for example because utils is already importing
sriov_config. This leads to duplicate code which is harder to maintain.

Change-Id: Ieafc2adc0247efc47130cd5ea29c59a140feaf0c