Fix PKEY_ID value in ifcfg file of infiniband child interface
In the current implementation of ifcfg file in os-net-config,
the high bit of the PKEY_ID setting is not being set, while
the high bit of the pkey in the DEVICE setting is being set.

However, due to a change in NetworkManager [1],
it is no longer necessary to set the high bit of the pkey in
the DEVICE setting, but applying this change in os-net-config
is causing compatibility issues with older versions of NetworkManager.

To address this issue, this patch proposes setting the high bit
for both the PKEY_ID and pkey in DEVICE settings in the ifcfg file.
This approach would work for all versions of NetworkManager
and maintain compatibility with older versions.

[1] https://github.com/NetworkManager/NetworkManager/commit/33f2f82a0951c79a5e0363413c63f9067fd85340

Closes-Bug: #2007372
Change-Id: I7b4f063508c4d9518ca706e62cf1368fba2d069e
(cherry picked from commit 3e73bd8c83325d11fbc4f2f13c7829ce07c2414f)